Initial implementation of StorageOverviewScreen

+++ core/api/src/main/kotlin/com/boswelja/truemanager/core/api/v2/pool/PoolV2ApiImpl.kt
@@ -1,4 +1,5 @@
 package com.boswelja.truemanager.core.api.v2.pool
+
 import io.ktor.client.HttpClient
 import io.ktor.client.call.body
 import io.ktor.client.request.get
@@ -261,7 +262,7 @@ internal data class StatsDto(
     @SerialName("self_healed")
     val selfHealed: Int,
     @SerialName("size")
-    val size: Int,
+    val size: Long,
     @SerialName("timestamp")
     val timestamp: Long,
     @SerialName("write_errors")
